# Optimize Date/Date32 scan