--- a/lib/column/datetime.go
+++ b/lib/column/datetime.go
@@ -302,9 +302,7 @@ func (col *DateTime) parseDateTime(value string) (tv time.Time, err error) {
 		return tv, nil
 	}
 	if tv, err = time.Parse(defaultDateTimeFormatNoZone, value); err == nil {
-		return time.Date(
-			tv.Year(), tv.Month(), tv.Day(), tv.Hour(), tv.Minute(), tv.Second(), tv.Nanosecond(), time.Local,
-		), nil
+		return getTimeWithDifferentLocation(tv, time.Local), nil
 	}
 	return time.Time{}, err
 }